refactor: move search view component to app router
