+++ src/app/(core)/teachers/search/Search.view.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from "next/router";
+import { useSearchParams } from "next/navigation";
 import { FC, useEffect } from "react";
 import {
   OakGrid,
@@ -15,13 +15,13 @@ import {
 import styled from "styled-components";
 import { useFeatureFlagVariantKey } from "posthog-js/react";
 
-import { SearchProps } from "./search.view.types";
+import { SearchProps } from "../../../../components/TeacherViews/Search/search.view.types";
 import {
   isKeyStageTitleValueType,
   removeHTMLTags,
   trackSearchModified,
-} from "./helpers";
-import { ContentFilterToggle } from "./ContentFilterToggle";
+} from "../../../../components/TeacherViews/Search/helpers";
+import { ContentFilterToggle } from "../../../../components/TeacherViews/Search/ContentFilterToggle";
 
 import { SearchResultsItemProps } from "@/components/TeacherComponents/SearchResultsItem";
 import useAnalytics from "@/context/Analytics/useAnalytics";
@@ -35,11 +35,11 @@ import NoSearchResults from "@/components/TeacherComponents/NoSearchResults";
 import {
   getActiveFilters,
   getSortedSearchFiltersSelected,
-} from "@/context/Search/search.helpers";
+} from "@/app/(core)/teachers/search/search.helpers";
 import SignPostToAila from "@/components/TeacherComponents/NoSearchResults/SignPostToAila";
 import MiniDropDown from "@/components/TeacherComponents/MiniDropdown";
 import SearchSuggestedFilters from "@/components/TeacherComponents/SearchSuggestedFilters/SearchSuggestedFilters";
-import { useSuggestedFilters } from "@/context/Search/useSuggestedFilters";
+import { useSuggestedFilters } from "@/app/(core)/teachers/search/useSuggestedFilters";
 import { SearchSuggestionBanner } from "@/components/TeacherComponents/SearchSuggestionBanner/SearchSuggestionBanner";
 import DelayedLoadingContainer from "@/components/SharedComponents/DelayedLoadingContainer";
 import { srOnlyCss } from "@/components/SharedComponents/ScreenReaderOnly";
@@ -87,7 +87,7 @@ const Search: FC<SearchProps> = (props) => {
   } = props;
   const { track } = useAnalytics();
   const { analyticsUseCase } = useAnalyticsPageProps();
-  const router = useRouter();
+  const searchParams = useSearchParams();
   const hitCount = results.length;
 
   const shouldShowError = status === "fail";
@@ -109,16 +109,17 @@ const Search: FC<SearchProps> = (props) => {
   }, [query.term, setSearchStartTime, status]);
 
   useEffect(() => {
+    const getParam = (param: string) => searchParams?.get(param) ?? "";
     const searchHasFilters =
-      router.query.keyStages ||
-      router.query.examBoards ||
-      router.query.contentTypes ||
-      router.query.subjects ||
-      router.query.yearGroups ||
-      router.query.curriculum;
+      getParam("keyStages") ||
+      getParam("examBoards") ||
+      getParam("contentTypes") ||
+      getParam("subjects") ||
+      getParam("yearGroups") ||
+      getParam("curriculum");
 
     if (
-      !router.query.page &&
+      !getParam("page") &&
       searchStartTime &&
       (status === "success" || status === "fail")
     ) {
@@ -145,7 +146,7 @@ const Search: FC<SearchProps> = (props) => {
           eventVersion: "2.0.0",
           analyticsUseCase: "Teacher",
           searchResultCount: hitCount,
-          activeFilters: getActiveFilters(router.query),
+          activeFilters: getActiveFilters(searchParams),
           searchTerm: query.term,
         });
       }
@@ -154,16 +155,15 @@ const Search: FC<SearchProps> = (props) => {
     analyticsUseCase,
     hitCount,
     query.term,
-    router,
+    searchParams,
     searchStartTime,
     setSearchStartTime,
     status,
     track,
   ]);
 
-  const searchFilterOptionSelected = getSortedSearchFiltersSelected(
-    router.query,
-  );
+  const searchFilterOptionSelected =
+    getSortedSearchFiltersSelected(searchParams);
 
   const searchResultExpanded = ({
     searchHit,
@@ -231,9 +231,8 @@ const Search: FC<SearchProps> = (props) => {
         unitSlug: searchHit.buttonLinkProps.unitSlug,
         analyticsUseCase: analyticsUseCase,
         searchRank: searchRank,
-        searchFilterOptionSelected: getSortedSearchFiltersSelected(
-          router.query,
-        ),
+        searchFilterOptionSelected:
+          getSortedSearchFiltersSelected(searchParams),
         searchResultCount: hitCount,
         searchResultType: searchHit.type,
         lessonName: removeHTMLTags(searchHit.title),
